requirements.txt
# === Data Manipulation ===
numpy==2.4.0

# === Data Handling ===
pyYAML==6.0.3

# === Market Data ===
yfinance==1.0

# === Market Data ===
langgraph==1.0.5
langchain==1.2.0
langchain-community==0.4.1
langchain-openai==1.1.6
tavily-python==0.7.17
pydantic==2.12.5
python-dotenv==1.2.1

-e .